Move CBOR serialization into silkworm_db
