+++ core/silkworm/types/receipt.hpp
@@ -29,10 +29,6 @@ struct Receipt {
     std::vector<Log> logs;
 };
 
-// TG-compatible CBOR encoding for storage.
-// See core/types/receipt.go
-Bytes cbor_encode(const std::vector<Receipt>& v);
-
 }  // namespace silkworm
 
 #endif  // SILKWORM_TYPES_RECEIPT_H_
